requirements.txt
streamlit
youtube-transcript-api
langchain
langchain-community
langchain-core
langchain-groq
faiss-cpu
sentence-transformers
python-dotenv
pydantic
numpy